+++ Source/Jeeves.Core/SecurityOption.cs
@@ -1,0 +1,9 @@
+﻿namespace Jeeves.Core
+{
+    public enum SecurityOption
+    {
+        Http,
+        Https,
+        HttpsAndAuthentication
+    }
+}